Add expansion argument for selective install

+++ install.py
@@ -41,6 +41,7 @@ def main():
     parser.add_argument('target', metavar=path.join('Path', 'To', 'Interface', 'AddOns'))
     parser.add_argument('-c', '--clean', action='store_true', help='clean targets before install')
     parser.add_argument('-u', '--uninstall', action='store_true', help='uninstall and exit')
+    parser.add_argument('-e', '--expansion', default=0, type=int, help='install single expansion')
     args = parser.parse_args()
 
     if not path.exists(args.target):
@@ -53,33 +54,34 @@ def main():
         sys.exit(1)
 
     for dir in sorted(os.listdir('plugins')):
-        plugin_dir = path.join('plugins', dir)
-        plugin_name = 'HandyNotes_' + dir[3:]
-        target_dir = path.join(target, plugin_name)
+        if args.expansion == 0 or args.expansion == int(dir[:2]):
+            plugin_dir = path.join('plugins', dir)
+            plugin_name = 'HandyNotes_' + dir[3:]
+            target_dir = path.join(target, plugin_name)
 
-        # remove previous installs
-        if args.clean or args.uninstall:
-            if path.exists(target_dir):
-                print(f'Deleting {target_dir}')
-                shutil.rmtree(target_dir)
-            if args.uninstall: continue
+            # remove previous installs
+            if args.clean or args.uninstall:
+                if path.exists(target_dir):
+                    print(f'Deleting {target_dir}')
+                    shutil.rmtree(target_dir)
+                if args.uninstall: continue
 
-        # install plugin-specific files
-        os.makedirs(target_dir, exist_ok=True)
-        for item in os.listdir(plugin_dir):
-            install(path.join(plugin_dir, item), path.join(target_dir, item))
+            # install plugin-specific files
+            os.makedirs(target_dir, exist_ok=True)
+            for item in os.listdir(plugin_dir):
+                install(path.join(plugin_dir, item), path.join(target_dir, item))
 
-        # add dummy .git directory so the plugins show up as "Working Copy" in the curse client
-        os.makedirs(path.join(target_dir, '.git'), exist_ok=True)
+            # add dummy .git directory so the plugins show up as "Working Copy" in the curse client
+            os.makedirs(path.join(target_dir, '.git'), exist_ok=True)
 
-        # link core + library dirs
-        for item in ('core', 'libs', 'embeds.xml'):
-            install(item, path.join(target_dir, item))
+            # link core + library dirs
+            for item in ('core', 'libs', 'embeds.xml'):
+                install(item, path.join(target_dir, item))
 
-        # install templates.xml file
-        with open('templates.xml') as f1, open(path.join(target_dir, 'templates.xml'), 'w') as f2:
-            print(f"Writing templates.xml => {path.join(target_dir, 'templates.xml')}")
-            f2.write(f1.read().format(addon=plugin_name))
+            # install templates.xml file
+            with open('templates.xml') as f1, open(path.join(target_dir, 'templates.xml'), 'w') as f2:
+                print(f"Writing templates.xml => {path.join(target_dir, 'templates.xml')}")
+                f2.write(f1.read().format(addon=plugin_name))
 
 if __name__ == '__main__':
     main()
